[Fix/#21] Response 타입 수정

+++ GLINT-iOS/GLINT-iOS/Source/Core/Utilities/FilterValueFormatter.swift
@@ -14,9 +14,9 @@ struct FilterValueFormatter {
     }
     
     // MARK: - 2. 조건부 포맷팅 (정수면 .0 제거)
-    static func formatSmart(_ value: Double) -> String {
+    static func formatSmart(_ value: Float) -> String {
         let rounded = (value * 10).rounded() / 10
-        if rounded == Double(Int(rounded)) {
+        if rounded == Float(Int(rounded)) {
             return String(Int(rounded))
         } else {
             return String(format: "%.1f", rounded)
@@ -25,8 +25,8 @@ struct FilterValueFormatter {
     
     static func photoMetaDataFormat(
         lensInfo: String?,
-        focalLength: Double?,
-        aperture: Double?,
+        focalLength: Float?,
+        aperture: Float?,
         iso: Int?
     ) -> String? {
         guard let lensInfo, let focalLength, let aperture, let iso else {
